+++ packages/seedcord/src/bot/decorators/Confirmable/flow.ts
@@ -3,7 +3,7 @@ import { ComponentType, type Message, type MessageComponentInteraction, type Mes
 import { isMessageComponentIx } from './utils';
 
 import type { ConfirmableOptions, ConfirmablePayload } from './types';
-import type { Repliables } from '@interfaces/Handler';
+import type { Repliables } from '@handlers/BaseHandler';
 
 export function shouldDefer<TComponentType extends MessageComponentType>(
     ix: Repliables,
