+++ packages/seedcord/src/customId/index.ts
@@ -1,3 +1,2 @@
 export { CustomId } from './CustomId';
 export { StaleCustomId, InvalidCustomId } from './Errors';
-export { ButtonHandler, ModalHandler, SelectHandler } from './Handlers';
