+++ packages/seedcord/src/handlers/event/index.ts
@@ -1,0 +1,2 @@
+export { EventHandler } from './EventHandler';
+export { EventMiddleware } from './EventMiddleware';
