+++ packages/seedcord/src/index.ts
@@ -15,6 +15,12 @@ export * from '@bUtilities/index';
 // Interfaces exports
 export * from '@interfaces/index';
 
+// Handlers exports
+export * from '@handlers/index';
+
+// Inputs exports
+export type * from '@inputs/index';
+
 // CustomId codec exports
 export * from '@customId/index';
