+++ packages/seedcord/src/inputs/AutocompleteOptions.ts
@@ -1,16 +1,7 @@
 import type { OptionKind, SlashOptionRegistry } from '@seedcord/types';
-import type { ApplicationCommandOptionChoiceData } from 'discord.js';
-import type { IsNever, Promisable } from 'type-fest';
+import type { IsNever } from 'type-fest';
 
 type Row<Route extends keyof SlashOptionRegistry> = SlashOptionRegistry[Route];
-
-// distributes over a union Route so a multi-command handler sees the union of every command's autocompletable
-// fields. keyof a union row would intersect to only the names common to all commands instead.
-type AutocompletableNames<Route extends keyof SlashOptionRegistry> = Route extends unknown
-    ? {
-          [Name in keyof Row<Route>]: Row<Route>[Name] extends { autocomplete: true } ? Name : never;
-      }[keyof Row<Route>]
-    : never;
 
 // the registry entry for one field name, gathered across every registered command that declares it.
 type EntryFor<Route extends keyof SlashOptionRegistry, Name extends PropertyKey> = Route extends unknown
@@ -18,30 +9,6 @@ type EntryFor<Route extends keyof SlashOptionRegistry, Name extends PropertyKey>
         ? Row<Route>[Name]
         : never
     : never;
-
-// the choice value Discord expects for an autocompletable option, string for a string option and number for
-// an integer or number option. distributes over a mixed-kind union to widen to string | number.
-type ChoiceValueOf<Entry> = Entry extends { kind: 'string' }
-    ? string
-    : Entry extends { kind: 'integer' | 'number' }
-      ? number
-      : never;
-
-/** The focused option. `value` is the raw partial Discord delivers mid-type, always a string regardless of kind. */
-export interface FocusedField<Route extends keyof SlashOptionRegistry> {
-    name: AutocompletableNames<Route>;
-    value: string;
-}
-
-/** One arm per autocompletable field, `value` is the focused partial, `respond` is pinned to the field's choice type. */
-export type FocusedArms<Route extends keyof SlashOptionRegistry, Ret> = {
-    [Name in AutocompletableNames<Route>]: (
-        value: string,
-        respond: (
-            choices: readonly ApplicationCommandOptionChoiceData<ChoiceValueOf<EntryFor<Route, Name>>>[]
-        ) => Promise<void>
-    ) => Promisable<Ret>;
-};
 
 // option names of a kind across a union Route, distributed to the union rather than the intersection.
 type NamesOfKind<Route extends keyof SlashOptionRegistry, Kind extends OptionKind> = Route extends unknown
