+++ packages/seedcord/src/inputs/index.ts
@@ -1,0 +1,2 @@
+export type { SlashOptions } from './SlashOptions';
+export type { AutocompleteOptions } from './AutocompleteOptions';
