+++ packages/seedcord/src/interfaces/index.ts
@@ -1,20 +1,5 @@
 export { BuilderComponent, CustomError, RowComponent, type RowType, type BuilderType } from './Components';
-export {
-    AutocompleteHandler,
-    EventHandler,
-    EventMiddleware,
-    InteractionHandler,
-    InteractionMiddleware,
-    type Repliables,
-    type ValidEventTypes,
-    type ValidInteractionTypes,
-    type ValidNonInteractionKeys,
-    type WithChecks
-} from './Handler';
 export { Pluggable, Plugin, type Initializeable, type PluginArgs, type PluginCtor } from './Plugin';
-export { SlashHandler } from './SlashHandler';
 
 // Type exports
-export type { AutocompleteOptions } from './AutocompleteOptions';
 export type * from './Core';
-export type { TypedOptions } from './TypedOptions';
