+++ packages/seedcord/tests/decorators/Confirmable.test.ts
@@ -2,8 +2,8 @@ import { ButtonInteraction, ComponentType, ContainerBuilder } from 'discord.js';
 import { beforeEach, describe, expect, it, vi, type Mock } from 'vitest';
 
 import { Confirmable } from '@bDecorators/Confirmable';
+import { InteractionHandler } from '@handlers/interaction';
 import { BuilderComponent } from '@interfaces/Components';
-import { InteractionHandler } from '@interfaces/Handler';
 
 import type { ConfirmableClassicOptions, ConfirmableOptions, RowLike } from '@bDecorators/Confirmable/types';
 import type { Core } from '@interfaces/Core';
