+++ packages/seedcord/tsconfig.json
@@ -10,6 +10,8 @@
             "@subscribers/*": ["./src/subscribers/*"],
             "@hmr/*": ["./src/hmr/*"],
             "@interfaces/*": ["./src/interfaces/*"],
+            "@handlers/*": ["./src/handlers/*"],
+            "@inputs/*": ["./src/inputs/*"],
             "@customId/*": ["./src/customId/*"],
             "@miscellaneous/*": ["./src/miscellaneous/*"],
