+++ packages/seedcord/vitest.config.ts
@@ -16,6 +16,8 @@ export default mergeConfig(
                 '@subscribers': path.resolve(__dirname, './src/subscribers'),
                 '@hmr': path.resolve(__dirname, './src/hmr'),
                 '@interfaces': path.resolve(__dirname, './src/interfaces'),
+                '@handlers': path.resolve(__dirname, './src/handlers'),
+                '@inputs': path.resolve(__dirname, './src/inputs'),
                 '@customId': path.resolve(__dirname, './src/customId'),
                 '@miscellaneous': path.resolve(__dirname, './src/miscellaneous'),
